Removed exports folder & moved exports into respective files
