+++ src/as/extras/OrbitController.ts
@@ -3,8 +3,8 @@ import { OrthographicCamera } from "../cameras/OrthographicCamera";
 import { PerspectiveCamera } from "../cameras/PerspectiveCamera";
 import { Event } from "../core/Event";
 import { Listener } from "../core/EventDispatcher";
-import { inputManager } from "../exports/io/InputManager";
-import { MouseEvent } from "../exports/io/MouseEvent";
+import { inputManager } from "./io/InputManager";
+import { MouseEvent } from "./io/MouseEvent";
 import { Matrix4 } from "../math/Matrix4";
 import { Quaternion } from "../math/Quaternion";
 import { Spherical } from "../math/Spherical";
