+++ src/as/extras/io/MouseEvent.ts
