+++ src/as/extras/ui/uiSignalManager.ts
