+++ src/as/geometries/index.ts
@@ -1,0 +1,4 @@
+export { createBox } from "./BoxGeometry";
+export {} from "./CapsuleGeometry";
+export { createPlane } from "./PlaneGeometry";
+export { createSphere } from "./SphereGeometry";
