+++ src/as/objects/routing/custom/Level1.ts
@@ -4,13 +4,13 @@ import { AmbientLight } from "../../../lights/AmbientLight";
 import { DirectionalLight } from "../../../lights/DirectionalLight";
 import { Color } from "../../../math/Color";
 import { Container } from "../core/Container";
-import { inputManager } from "../../../exports/io/InputManager";
-import { KeyboardEvent } from "../../../exports/io/KeyboardEvent";
+import { inputManager } from "../../../extras/io/InputManager";
+import { KeyboardEvent } from "../../../extras/io/KeyboardEvent";
 import { Listener } from "../../../core/EventDispatcher";
-import { uiSignaller } from "../../../exports/ui/uiSignalManager";
+import { uiSignaller } from "../../../extras/ui/uiSignalManager";
 import { Event } from "../../../core/Event";
 import { UIEventType } from "../../../../common/UIEventType";
-import { UIEvent } from "../../../exports/ui/UIEvent";
+import { UIEvent } from "../../../extras/ui/UIEvent";
 import { Link } from "../core/Link";
 import { TransformNode } from "../../../core/TransformNode";
 import { lock, unlock } from "../../../Imports";
